refactor: seperate API module & code style
