+++ client/modules/color.py
@@ -1,0 +1,2 @@
+def Green(msg:str) -> str:
+    return f'\033[32m{msg}\033[0m'
